+++ sphinx/screens/chats/chat-common/chat-common/src/main/java/chat/sphinx/chat_common/ui/ChatViewModel.kt
@@ -16,7 +16,6 @@ import androidx.annotation.CallSuper
 import androidx.core.view.inputmethod.InputConnectionCompat
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavArgs
 import app.cash.exhaustive.Exhaustive
@@ -35,15 +34,16 @@ import chat.sphinx.chat_common.ui.viewstate.attachment.AttachmentSendViewState
 import chat.sphinx.chat_common.ui.viewstate.footer.FooterViewState
 import chat.sphinx.chat_common.ui.viewstate.header.ChatHeaderViewState
 import chat.sphinx.chat_common.ui.viewstate.menu.ChatMenuViewState
+import chat.sphinx.chat_common.ui.viewstate.menu.MoreMenuOptionsViewState
+import chat.sphinx.chat_common.ui.viewstate.messageholder.*
 import chat.sphinx.chat_common.ui.viewstate.messageholder.BubbleBackground
-import chat.sphinx.chat_common.ui.viewstate.messageholder.InvoiceLinesHolderViewState
 import chat.sphinx.chat_common.ui.viewstate.messageholder.LayoutState
 import chat.sphinx.chat_common.ui.viewstate.messageholder.MessageHolderViewState
 import chat.sphinx.chat_common.ui.viewstate.messagereply.MessageReplyViewState
+import chat.sphinx.chat_common.ui.viewstate.search.MessagesSearchViewState
 import chat.sphinx.chat_common.ui.viewstate.selected.SelectedMessageViewState
 import chat.sphinx.chat_common.util.*
 import chat.sphinx.concept_image_loader.ImageLoaderOptions
-import chat.sphinx.concept_image_loader.Transformation
 import chat.sphinx.concept_link_preview.LinkPreviewHandler
 import chat.sphinx.concept_link_preview.model.TribePreviewName
 import chat.sphinx.concept_link_preview.model.toPreviewImageUrlOrNull
@@ -66,7 +66,6 @@ import chat.sphinx.wrapper_common.*
 import chat.sphinx.wrapper_common.chat.ChatUUID
 import chat.sphinx.wrapper_common.dashboard.ChatId
 import chat.sphinx.wrapper_common.dashboard.ContactId
-import chat.sphinx.wrapper_common.feed.FeedId
 import chat.sphinx.wrapper_common.lightning.*
 import chat.sphinx.wrapper_common.message.MessageId
 import chat.sphinx.wrapper_common.message.MessageUUID
@@ -97,7 +96,6 @@ import org.jitsi.meet.sdk.JitsiMeetActivity
 import org.jitsi.meet.sdk.JitsiMeetConferenceOptions
 import org.jitsi.meet.sdk.JitsiMeetUserInfo
 import java.io.*
-import java.util.*
 import kotlin.collections.ArrayList
 
 
@@ -144,12 +142,19 @@ abstract class ChatViewModel<ARGS: NavArgs>(
             .build()
     }
 
+    val messagesSearchViewStateContainer: ViewStateContainer<MessagesSearchViewState> by lazy {
+        ViewStateContainer(MessagesSearchViewState.Idle)
+    }
 
     val messageReplyViewStateContainer: ViewStateContainer<MessageReplyViewState> by lazy {
         ViewStateContainer(MessageReplyViewState.ReplyingDismissed)
     }
 
     val callMenuHandler: ViewStateContainer<MenuBottomViewState> by lazy {
+        ViewStateContainer(MenuBottomViewState.Closed)
+    }
+
+    val moreOptionsMenuHandler: ViewStateContainer<MenuBottomViewState> by lazy {
         ViewStateContainer(MenuBottomViewState.Closed)
     }
 
@@ -220,7 +225,7 @@ abstract class ChatViewModel<ARGS: NavArgs>(
         ChatHeaderViewStateContainer()
     }
 
-    protected suspend fun getChat(): Chat {
+    suspend fun getChat(): Chat {
         chatSharedFlow.replayCache.firstOrNull()?.let { chat ->
             return chat
         }
@@ -272,7 +277,7 @@ abstract class ChatViewModel<ARGS: NavArgs>(
         owner: Contact
     ): InitialHolderViewState
 
-    private suspend fun getOwner(): Contact {
+    suspend fun getOwner(): Contact {
         return contactRepository.accountOwner.value.let { contact ->
             if (contact != null) {
                 contact
@@ -350,6 +355,8 @@ abstract class ChatViewModel<ARGS: NavArgs>(
             contactRepository.getContactByPubKey(it).firstOrNull()
         } ?: null
 
+        var unseenSeparatorAdded = false
+
         val newList = ArrayList<MessageHolderViewState>(messages.size)
 
         withContext(io) {
@@ -387,6 +394,37 @@ abstract class ChatViewModel<ARGS: NavArgs>(
                     openReceivedPaidInvoicesCount > 0
                 )
 
+                if (!message.seen.isTrue() && !sent && !unseenSeparatorAdded) {
+                    newList.add(
+                        MessageHolderViewState.Separator(
+                            MessageHolderType.UnseenSeparator,
+                            null,
+                            chat,
+                            tribeAdmin,
+                            BubbleBackground.Gone(setSpacingEqual = true),
+                            invoiceLinesHolderViewState,
+                            InitialHolderViewState.None,
+                            accountOwner = { owner }
+                        )
+                    )
+                    unseenSeparatorAdded = true
+                }
+
+                if (previousMessage == null || message.date.isDifferentDayThan(previousMessage.date)) {
+                    newList.add(
+                        MessageHolderViewState.Separator(
+                            MessageHolderType.DateSeparator,
+                            message.date,
+                            chat,
+                            tribeAdmin,
+                            BubbleBackground.Gone(setSpacingEqual = true),
+                            invoiceLinesHolderViewState,
+                            InitialHolderViewState.None,
+                            accountOwner = { owner }
+                        )
+                    )
+                }
+
                 val isDeleted = message.status.isDeleted()
 
                 if (
@@ -414,6 +452,7 @@ abstract class ChatViewModel<ARGS: NavArgs>(
                                 }
                             },
                             invoiceLinesHolderViewState = invoiceLinesHolderViewState,
+                            highlightedText = null,
                             messageSenderInfo = { messageCallback ->
                                 when {
                                     messageCallback.sender == chat.contactIds.firstOrNull() -> {
@@ -484,6 +523,7 @@ abstract class ChatViewModel<ARGS: NavArgs>(
                                     getInitialHolderViewStateForReceivedMessage(message, owner)
                                 }
                             },
+                            highlightedText = null,
                             messageSenderInfo = { messageCallback ->
                                 when {
                                     messageCallback.sender == chat.contactIds.firstOrNull() -> {
@@ -750,8 +790,9 @@ abstract class ChatViewModel<ARGS: NavArgs>(
 
     protected abstract fun forceKeyExchange()
 
+    var messagesLoadJob: Job? = null
     fun screenInit() {
-        viewModelScope.launch(mainImmediate) {
+        messagesLoadJob = viewModelScope.launch(mainImmediate) {
             messageRepository.getAllMessagesToShowByChatId(getChat().id, 20).firstOrNull()?.let { messages ->
                 messageHolderViewStateFlow.value =
                     getMessageHolderViewStateList(messages).toList()
@@ -892,6 +933,71 @@ abstract class ChatViewModel<ARGS: NavArgs>(
                     }
                 }
             }
+        }
+    }
+
+    var messagesSearchJob: Job? = null
+    suspend fun searchMessages(text: String?) {
+        moreOptionsMenuHandler.updateViewState(
+            MenuBottomViewState.Closed
+        )
+
+        if (messagesSearchViewStateContainer.viewStateFlow.value is MessagesSearchViewState.Idle) {
+            loadAllMessages()
+        }
+
+        chatId?.let { nnChatId ->
+            text?.let { nnText ->
+                if (nnText.toCharArray().size > 2) {
+                    messagesSearchViewStateContainer.updateViewState(
+                        MessagesSearchViewState.Loading
+                    )
+
+                    messagesSearchJob?.cancel()
+                    messagesSearchJob = viewModelScope.launch(io) {
+                        delay(500L)
+
+                        messageRepository.searchMessagesBy(nnChatId, nnText).firstOrNull()?.let { messages ->
+                            messagesSearchViewStateContainer.updateViewState(
+                                MessagesSearchViewState.Searching(messages, 0, true)
+                            )
+                        }
+                    }
+                    return
+                }
+            }
+        }
+
+        messagesSearchViewStateContainer.updateViewState(
+            MessagesSearchViewState.Searching(emptyList(), 0, true)
+        )
+    }
+
+    fun navigateResults(
+        advanceBy: Int
+    ) {
+        val searchViewState = messagesSearchViewStateContainer.viewStateFlow.value
+        if (searchViewState is MessagesSearchViewState.Searching) {
+            messagesSearchViewStateContainer.updateViewState(
+                MessagesSearchViewState.Searching(
+                    searchViewState.messages,
+                    searchViewState.index + advanceBy,
+                    advanceBy > 0
+                )
+            )
+        }
+    }
+
+    private fun loadAllMessages() {
+        if (messagesLoadJob?.isActive == true) {
+            messagesLoadJob?.cancel()
+        }
+        messagesLoadJob = viewModelScope.launch(io) {
+            messageRepository.getAllMessagesToShowByChatId(getChat().id, 0)
+                .distinctUntilChanged().collect { messages ->
+                    messageHolderViewStateFlow.value =
+                        getMessageHolderViewStateList(messages).toList()
+                }
         }
     }